+++ src/SingleEdit/ResizeStaff.js
