+++ src/SingleEdit/SelectOptions.js
@@ -1,8 +1,8 @@
 /** @module SelectOptions */
-import * as Contents from './Contents.js';
+import * as Contents from '../utils/Contents.js';
 import * as Grouping from './Grouping.js';
-import * as Notification from './Notification.js';
-import InfoBox from './InfoBox.js';
+import * as Notification from '../Notification.js';
+import InfoModule from '../InfoModule.js';
 import SplitHandler from './SplitHandler.js';
 const $ = require('jquery');
 
@@ -130,7 +130,7 @@ export function triggerNeumeActions () {
   }
 
   $('.grouping').on('click', (e) => {
-    var contour = InfoBox.getContourByValue(e.target.id);
+    var contour = InfoModule.getContourByValue(e.target.id);
     triggerChangeGroup(contour);
   });
