+++ src/SingleEdit/SingleEditMode.js
@@ -1,5 +1,5 @@
-import * as Controls from './Controls.js';
-import DragHandler from './DragHandler.js';
+import * as Controls from '../utils/Controls.js';
+import DragHandler from '../SingleView/DragHandler.js';
 import * as Select from './Select.js';
 import InsertHandler from './InsertHandler.js';
 import * as SelectOptions from './SelectOptions.js';
