+++ src/SingleEdit/SplitHandler.js
@@ -1,4 +1,4 @@
-import * as Notification from './Notification.js';
+import * as Notification from '../Notification.js';
 const $ = require('jquery');
 
 /**
