+++ src/SingleView/DragHandler.js
